manage books component
